requirements.txt
psutil>=5.9.0  # Process management
pyyaml>=6.0    # Already exists
ultralytics>=8.0.0  # YOLO training